+++ src/main/java/com/beautiflow/chat/domain/ChatRoom.java
@@ -1,19 +1,20 @@
-package com.beautiflow.chat;
+package com.beautiflow.chat.domain;
 
-import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
+import com.beautiflow.chat.domain.ChatMessage;
 import com.beautiflow.global.domain.BaseTimeEntity;
-import com.beautiflow.global.domain.SenderType;
+import com.beautiflow.shop.domain.Shop;
 import com.beautiflow.user.domain.User;
 
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -26,21 +27,21 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 @Entity
-@Table(name = "messages")
-public class ChatMessage extends BaseTimeEntity {
+@Table(name = "chat_rooms")
+public class ChatRoom extends BaseTimeEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
 	@ManyToOne(fetch = FetchType.LAZY)
-	private ChatRoom chatRoom;
+	private Shop shop;
 
 	@ManyToOne(fetch = FetchType.LAZY)
-	private User sender;
+	private User customer;
 
-	@Enumerated(EnumType.STRING)
-	private SenderType senderType;
+	@ManyToOne(fetch = FetchType.LAZY)
+	private User designer;
 
-	private String content;
-	private String imageUrl;
+	@OneToMany(mappedBy = "chatRoom")
+	private List<ChatMessage> messages = new ArrayList<>();
 }
